requirements.txt
groq>=0.9.0
google-auth>=2.0.0
google-auth-oauthlib>=1.1.0
google-api-python-client>=2.100.0
gspread>=6.0.0
